+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## Unreleased
+## v0.17.0 (15 Jan. 2022)
 ### Added
 - Add `RigidBody::set_enabled`, `RigidBody::is_enabled`, `RigidBodyBuilder::enabled` to enable/disable a rigid-body
   without having to delete it. Disabling a rigid-body attached to a multibody joint isn’t supported yet.
@@ -11,6 +11,7 @@
   a port of the vehicle controller from Bullet physics).
 - Add `RigidBody::user_force` and `RigidBody::user_torque` to read the forces or torques added by the user to a
   dynamic rigid-body.
+- Add `RigidBody::locked_axes` to get the rigid-body axes that were locked by the user.
 
 ### Modified
 - Add the `QueryPipeline` as an optional argument to `PhysicsPipeline::step` and `CollisionPipeline::step`. If this
@@ -18,6 +19,7 @@
   these other pipelines. In that case, calling `QueryPipeline::update` a `PhysicsPipeline::step` isn’t needed.
 - `RigidBody::set_body_type` now takes an extra boolean argument indicating if the rigid-body should be woken-up
   (if it becomes dynamic).
+- `RigidBody::mass_properties` now also returns the world-space mass-properties of the rigid-body.
 
 ### Fix
 - Fix bug resulting in rigid-bodies being awakened after they are created, even if they are created sleeping.